fixed end data in schedules and toggle update in staff profile and some ui issues
