+++ package.json
@@ -13,7 +13,7 @@
     "@expo/vector-icons": "^14.0.3",
     "@react-native-community/checkbox": "^0.5.17",
     "@react-native-community/datetimepicker": "8.0.1",
-    "@react-native-masked-view/masked-view": "^0.3.2",
+    "@react-native-masked-view/masked-view": "0.3.1",
     "@react-native-picker/picker": "2.7.5",
     "@react-navigation/drawer": "^6.7.2",
     "@react-navigation/elements": "^1.3.31",
@@ -67,8 +67,8 @@
     "react-native-popover-view": "^5.1.8",
     "react-native-pure-chart": "^0.0.24",
     "react-native-reanimated": "~3.10.1",
-    "react-native-safe-area-context": "^4.14.1",
-    "react-native-screens": "^3.35.0",
+    "react-native-safe-area-context": "4.10.5",
+    "react-native-screens": "3.31.1",
     "react-native-svg": "15.2.0",
     "react-native-tab-view": "^4.0.5",
     "react-native-table-component": "^1.2.2",
